Trying to fix it for mac
For some reason, it didn't want to get all of the windows opened on mac. So I switched it to track window that user is going to click on after starting the app.

Still having some issues with multi windowed apps on Mac. Got reports of not counting time correctly while running for longer period of time.